Assignment: You are in charge of designing the denominations of coins. However, we're going to do this for old English pounds which were worth 240 pence. You have made a study of the subpound prices and have determined that each multiple of 5 cents price is N times as likely as a non-multiple of 5 pence. For example, if N = 4, then 15 pence is four times as likely to be the price as 43 pence. But any 5 pence multiple is equally likely as any other 5 pence multiple and ditto for the non-5 pence multiples. The N will be given in class and your programs will have 2 minutes to solve the following problem: (N will be 1 or greater but may not be an integer.)
Your first job is to design a set of 4 coin denominations such that
the expected number of coins required to give
exact change for a purchase is minimized
given these constraints.
This is called the Exact Change Number.
Using the U.S. denominations, the Exact Change Number for
43 pence can be realized by one quarter, one dime, one nickel,
and three pennies, thus giving a total of 6.